requirements.txt
fastapi
uvicorn[standard]
redis
pyjwt
jinja2
python-dotenv
pytest
asyncio